debian: ensure updated files get installed
The existing `userChrome.css` checks the "last modified" date in order
to check if it should be updated or not. For some reason, this date
isn't updated when installing a new version of the package, causing the
old version to be used for existing installs. Adding a `postinst` script
allows us to work around this issue by using `touch` so the "last
modified" timestamp is updated.