userChrome/tabmenu: Fix height of tab menu in FF 113 (MR 31)
Without this fix, the tab menu is only tall enough for 2 or 3 entries.

This is an update of the rule directly above it, changing the ID and
changing from a calculated height to a fixed height matching its child
#allTabsMenu-allTabsView. With the calculated height, the menu entries
were offset above the top of the menu. The previous rule no longer seems
to be necessary, at least in Firefox 102 and newer, since
#allTabsMenu-multiView does not exist.