editorconfig: fix syntax
Fix the pattern not working in neovim's now built-in editorconfig
support by removing "**/". This should have matched all paths, but it
doesn't work with the built-in parser and by removing it, we apparently
match all paths too. This fixes neovim indenting with tabs in the
Makefile and shell scripts in this repository.

Happy to see that this is a built-in feature now \o/