menupopup: move if we should be pinned to the bottom or if we don't fit
Previously, the keyboard's appearance could cause menus to overflow out of
the screen, which would cause options to be unreachable. By backporting the old
resize logic from panels and adapting it a little, we can scoot menus out of harm's way.

This ensures menus are always usable even if the keyboard appears/disappears.